Extract Keysight error handling into mixin class
