+++ qcodes/instrument_drivers/Keysight/Keysight_33500B_channels.py
@@ -5,6 +5,9 @@ from functools import partial
 from qcodes.instrument.channel import InstrumentChannel, ChannelList
 import logging
 
+from qcodes.instrument_drivers.Keysight.private.error_handling import \
+    KeysightErrorQueueMixin
+
 log = logging.getLogger(__name__)
 
 
@@ -230,7 +233,7 @@ class KeysightChannel(InstrumentChannel):
                            )
 
 
-class Keysight_33500B_Channels(VisaInstrument):
+class Keysight_33500B_Channels(KeysightErrorQueueMixin, VisaInstrument):
     """
     QCoDeS driver for the Keysight 33500B Waveform Generator using QCoDeS
     channels
@@ -252,21 +255,6 @@ class Keysight_33500B_Channels(VisaInstrument):
                       " instead.", UserWarning)
 
         super().__init__(name, address, **kwargs)
-
-        def errorparser(rawmssg):
-            """
-            Parses the error message.
-
-            Args:
-                rawmssg (str): The raw return value of 'SYSTem:ERRor?'
-
-            Returns:
-                tuple (int, str): The error code and the error message.
-            """
-            code = int(rawmssg.split(',')[0])
-            mssg = rawmssg.split(',')[1].strip().replace('"', '')
-
-            return code, mssg
 
         channels = ChannelList(self, "Channels", KeysightChannel,
                                snapshotable=False)
@@ -292,39 +280,9 @@ class Keysight_33500B_Channels(VisaInstrument):
                            vals=vals.Enum('ON', 'OFF')
                            )
 
-        self.add_parameter('error',
-                           label='Error message',
-                           get_cmd='SYSTem:ERRor?',
-                           get_parser=errorparser
-                           )
-
         self.add_function('force_trigger', call_cmd='*TRG')
 
         self.add_function('sync_channel_phases', call_cmd='PHAS:SYNC')
 
         if not silent:
             self.connect_message()
-
-    def flush_error_queue(self, verbose=True):
-        """
-        Clear the instrument error queue.
-
-        Args:
-            verbose (Optional[bool]): If true, the error messages are printed.
-                Default: True.
-        """
-
-        log.debug('Flushing error queue...')
-
-        err_code, err_message = self.error()
-        log.debug('    {}, {}'.format(err_code, err_message))
-        if verbose:
-            print(err_code, err_message)
-
-        while err_code != 0:
-            err_code, err_message = self.error()
-            log.debug('    {}, {}'.format(err_code, err_message))
-            if verbose:
-                print(err_code, err_message)
-
-        log.debug('...flushing complete')
